add - delete cover image from cloudinary before deleting post

+++ src/app/actions/deletePost.ts
@@ -4,8 +4,7 @@ import { auth } from '@clerk/nextjs/server'
 import { neon } from '@neondatabase/serverless'
 import { revalidatePath } from 'next/cache'
 import { redirect } from 'next/navigation'
-
-// TODO before deleting post, find post by slug and delete the image from Cloudinary (by calling server action deleteImageFromCloudinary and passing the cover_image_cloudinary, which is the public_id of the image)
+import { deleteImageFromCloudinary } from './deleteImageFromCloudinary'
 
 // Delete a post from the database
 export async function deletePost(slug: string) {
@@ -13,21 +12,39 @@ export async function deletePost(slug: string) {
     const accessMessage = await checkUserAccess()
     if (accessMessage) return { message: accessMessage }
 
-    // * Delete the post from the database
     const sql = neon(process.env.DATABASE_URL as string)
+
+    // * Get the cover image public_id
+    const post = await sql`
+      SELECT cover_image_cloudinary
+      FROM posts
+      WHERE slug = ${slug}
+    `
+
+    if (!post || !post.length) {
+      throw new Error('Errore - Post non trovato')
+    }
+
+    // * Delete the cover image from Cloudinary if it exists
+    const coverImagePublicId = post[0]?.cover_image_cloudinary
+    if (coverImagePublicId) {
+      await deleteImageFromCloudinary(coverImagePublicId)
+    }
+
+    // * Delete the post from the database
     await sql`
       DELETE FROM posts
       WHERE slug = ${slug}
     `
-
-    revalidatePath('/') // Revalidate the home page
-    redirect('/') // Redirect home
   } catch (error) {
     console.error('Error deleting post:', error)
     return {
       message: 'Errore interno del server - Impossibile eliminare il post',
     }
   }
+
+  revalidatePath('/') // Revalidate the home page
+  redirect('/') // Redirect home
 }
 
 // Check if the user has access to publish a post
